+++ batch-server/src/main/java/com/pickmylunch/batch/pipeline/util/dto/AddressDto.java
@@ -1,4 +1,4 @@
-package com.pickmylunch.batch.pipeline.dto;
+package com.pickmylunch.batch.pipeline.util.dto;
 
 public record AddressDto(String dosi, String sigungu, String detail) {
 }
